pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "virosense"
version = "0.1.0"
description = "Multi-modal viral detection and characterization using DNA and protein structural analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["bioinformatics", "virology", "metagenomics", "evo2", "viral detection"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "click>=8.0",
    "biopython>=1.80",
    "numpy>=1.24",
    "pandas>=2.0",
    "loguru>=0.7",
    "requests>=2.28",
    "tqdm>=4.60",
    "scikit-learn>=1.3",
    "h5py>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gpu = ["torch>=2.0", "evo2"]
vhold = ["torch>=2.0", "transformers>=4.30", "sentencepiece>=0.1.99"]
nim = ["httpx>=0.25"]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]

[project.scripts]
virosense = "virosense.cli:main"

[project.urls]
Repository = "https://github.com/virosense/virosense"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "integration: requires NVIDIA_API_KEY",
    "e2e: end-to-end tests requiring GPU or NIM",
]
